requirements.txt
arrow==0.13.2	
astroid==2.2.5	
awscli==1.18.53	
beautifulsoup4==4.8.2	
boto3==1.9.240	
botocore==1.16.3	
certifi==2019.3.9	
chardet==3.0.4	
colorama==0.4.1	
dateparser==0.7.1	
Django==2.2.2	
django-appconf==1.0.4	
django-classy-tags==1.0.0	
django-cms==3.7.2	
django-cors-headers==3.2.0	
django-filer==1.7.1	
django-filter==2.3.0	
django-formtools==2.2	
django-js-asset==1.2.2	
django-modelcluster==5.0.2	
django-mptt==0.11.0	
django-paypal==1.0.0	
django-pdb==0.6.2	
django-polymorphic==2.1.2	
django-sekizai==1.1.0	
django-smtp-ssl==1.0	
django-storages==1.7.2	
django-taggit==1.3.0	
django-tinymce==3.0.2	
django-treebeard==4.3	
djangocms-admin-style==1.5.0	
djangocms-attributes-field==1.2.0	
djangocms-column==1.9.0	
djangocms-file==2.4.0	
djangocms-link==2.6.0	
djangocms-picture==2.4.0	
djangocms-snippet==2.3.0	
djangocms-style==2.3.0	
djangocms-text-ckeditor==3.9.0	
djangocms-video==2.3.0	
djangorestframework==3.11.1	
docutils==0.15.2	
draftjs-exporter==2.1.7	
easy-thumbnails==2.7	
et-xmlfile==1.0.1	
ffmpeg-python==0.2.0	
future==0.18.2	
html5lib==1.1	
idna==2.10	
iexfinance==0.4.2	
isort==4.3.21	
jdcal==1.4.1	
jmespath==0.9.4	
jsonfield==3.1.0	
l18n==2018.5	
lazy-object-proxy==1.4.1	
libmagic==1.0	
lxml==4.3.3	
mccabe==0.6.1	
moment==0.8.2	
mysqlclient==2.0.1	
numpy==1.16.3	
opencv-python-headless==4.1.1.26	
openpyxl==3.0.4	
pandas==0.24.2	
pandas-datareader==0.7.0	
Pillow==7.2.0	
pyasn1==0.4.8	
pylint==2.3.1	
python-dateutil==2.8.0	
python-magic==0.4.15	
pytz==2019.1	
PyYAML==5.3.1	
regex==2019.5.25	
requests==2.24.0	
rsa==3.4.2	
s3transfer==0.3.3	
six==1.12.0	
soupsieve==2.0.1	
sqlparse==0.3.0	
tablib==2.0.0	
times==0.7	
typed-ast==1.4.0	
tzlocal==1.5.1	
Unidecode==1.1.1	
urllib3==1.25.3	
wagtail==2.10	
webencodings==0.5.1	
Willow==1.4	
wrapt==1.11.1	
xlrd==1.2.0	
XlsxWriter==1.3.2	
xlwt==1.3.0
django-smtp-ssl==1.0
mailchimp-marketing==3.0.72